Add files via upload
The Kruskal-Wallis test is a non-parametric statistical test used to determine whether there are any statistically significant differences between the medians of three or more independent groups. It is an extension of the Mann-Whitney U test (or Wilcoxon rank-sum test) for more than two groups.

This line of code performs the Kruskal-Wallis test on the variable Sales grouped by the variable Design. The formula Sales ~ Design specifies that we want to test the hypothesis that the distribution of sales is the same across different levels of the design variable. The kruskal.test function is used for conducting the Kruskal-Wallis test.

Explain the Output:
The output of the kruskal.test function will include the test statistic and the p-value. If the p-value is below a chosen significance level (e.g., 0.05), you may reject the null hypothesis and conclude that there are significant differences in the distribution of sales across different design levels.