Add files via upload
ANCOVA is a statistical technique that combines elements of analysis of variance (ANOVA) and regression. It is used to compare means of dependent variables across levels of a categorical independent variable while statistically controlling for the effects of other continuous variables, known as covariates. The goal is to determine if there are any statistically significant differences between the group means after accounting for the influence of covariates.